ci: remove docs workflow
This workflow originated from a custom repository template of mine, `zerotask-rust-lib-template`. For libraries it makes sense to have an up-to-date version of the Rust docs. But in this case, there is no value added by building the documentation on each push to master.

Alongside this, the `gh-pages` branch will also be deleted.