feat!: allow further editing after questioning process (#56)
BREAKING CHANGE: The `-e,--edit` flag behaves differently. Instead of opening
the editor directly at the start the CLI will now first question the user for
all answers and open the commit message afterwards in the editor. This can be
used to further edit or review the message before finally commiting it.

Closes #55